move some code in separated objects
